Refactor and README WIP
